new-relic-exporter-base/requirements.txt
opentelemetry.api
opentelemetry.sdk
python-gitlab
opentelemetry.exporter.otlp.proto.grpc
opentelemetry.instrumentation.logging
pyrfc3339
zulu
schedule
regex